Move structs that are only used in mem_impl
